Fixed a bug in the regression test to use the right filenames for each platform; it was causing the test to fail on Windows.
Added some documentation to bin/update-release-info.py.

In my last log message, I forgot to mention that QMTest/TestRuntest.py was
tweaked to make it less specialized for testing runtest.py.  It can now be
used to test other programs; in this case, bin/update-release-info.py.  The
defaults are unchanged so that it will normally test runtest.py.